Rewrite Apple PAL using native APIs (#336)
This is a rewrite of the Apple PAL that implements the AlignedAllocation, Entropy, and LazyCommit features through native Apple APIs.

It adds a dependency on Security.framework via SecRandomCopyBytes. Apple actively discourages use of getentropy and the symbol is not allowed on the App Store.

+++ src/pal/pal_consts.h
@@ -65,7 +65,7 @@ namespace snmalloc
   /**
    * Default Tag ID for the Apple class
    */
-  static const int PALAnonDefaultID = 241;
+  static const uint8_t PALAnonDefaultID = 241;
 
   /**
    * This struct is used to represent callbacks for notification from the
